Extract native repository review endpoints from bridge
