+++ packages/agent-native-bridge/agent-native-bridge.cabal
@@ -74,6 +74,7 @@ foreign-library haskell-agent-bridge
                     , Agent.CLI.MacOS.RepositoryWorkers
                     , Agent.CLI.MacOS.RepositoryChecks
                     , Agent.CLI.MacOS.RepositoryDeliveryBridge
+                    , Agent.CLI.MacOS.RepositoryReviewBridge
                     , Agent.CLI.MacOS.RepositoryInput
                     , Agent.CLI.MacOS.ResourceAdmin
     c-sources:        cbits/HaskellAgentBridgeRuntime.c
@@ -157,6 +158,7 @@ test-suite agent-native-bridge-test
                      , Agent.CLI.MacOS.RepositoryWorkers
                      , Agent.CLI.MacOS.RepositoryChecks
                      , Agent.CLI.MacOS.RepositoryDeliveryBridge
+                     , Agent.CLI.MacOS.RepositoryReviewBridge
                     , Agent.CLI.MacOS.ResourceAdmin
                      , Agent.CLI.MacOS.BrowserBridgeFFISpec
                      , Agent.CLI.MacOS.ComputerBridgeSpec
